Config options for cmdbuf location/size.
Will default to putting the cmdbuf in AGP (except for PowerPC) if an AGP card
is detected.  Won't currently be detected because something is wrong DRM-size.

If you have an AGP card you can manually specify 'FBLocation "AGP"' in
xorg.conf to make it work.

Needs an updated DRM to have any effect.