nv50: tile offscreen pixmaps
Until someone figures out how to make the display engine scan out from
tiled surfaces things will be a bit painful.  Future commits for exa
composite and Xv will require the use of the 3D engine, which will *not*
render to linear surfaces.. a composite manager will probably be a must
for it to be usable.

I really don't want the NOUVEAU_MEM_TILE ioctl in the end, but we'll have
to deal with it until we can drop support for <xorg-server-1.5.  EXA driver
handled pixmaps are far far nicer :)