nv50: Xv support
The NV12 YCbCr->RGB shader is hand-written just like the composite shaders,
hopefully correct!  I'll probably put a disasm next to each opcode at some
point so it doesn't look like just a bunch of magics.

You'll need to use a composite manager if you want to watch videos, we can't
render directly to the screen for the same reasons as mentioned in the
composite commit.  Though, a simple workaround is possible in the Xv case,
so that'll probably come at some point.  For now, PutImage() will fail with
BadMatch.

The Y and UV plane upload may not be 100% for what NV50 requires, I want to
rework all that at some point anyway.  For now, at least there's some form
of accelerated video to work with.

Enjoy :)