randr12: multiple encoders per connector (DVI-I)
that this doesn't break things in subtle and unexpected ways is too much to
hope for. one not-subtle and expected way things will break is that xorg.confs
with specific setups for DVI-A-n and DVI-D-n connections will need renaming (to
DVI-I-n).

as the last patch in the series, this also fixes an issue introduced with
144b04a2 (randr12: fix dpms, detect, destroy, ...), that for a screen connected
to a dvi connector, both the analogue and digital variants were detected as
connected