Revert "nv50: make entire offscreen area tiled, use extra blits to scanout buffer"
This reverts commit 6c3f8da1bff3c94353748321b40a523add824327.

EXA doesn't use UTS/DFS hooks when it's accessing the "front buffer" for
software fallbacks, so sw rendering to it was garbled because it doesn't
understand tiling.

The Prepare/FinishAccess hooks could be used to point the pixmap at the
linear scanout buffer instead for sw access, but on VT switch for some
reason exa get called to try and do things, hit the PrepareAccess() hook
and die horribly because in LeaveVT() the driver deallocates the buffers.

The other alternative was to use libwfb to wrap the fb accesses and
handle tiling, but exa isn't quite ready for that yet either - it's
being worked on.

Another point made about this patch was that we should figure out the
display engine and get it to scanout the tiled buffer directly, not
commit shoddy workarounds.  Even for this "correct" solution, the two
above-mentioned issues still apply.