kms: work around some xserver stupidity with DPMS
The server calls the output DPMS functions with the intention of
turning off outputs before reprogramming them.  It probably shouldn't
be doing this in the case of the randr driver using set_mode_major,
and leave it up to the driver.

The DRM modesetting will usually reenable the outputs after a modeset
anyway, which is the behaviour that the xserver currently relies on.
However, there's a shortcut used in the drm to allow far faster fb
swaps without doing a modeset.  When this path is hit, the DRM will
rightly not touch any power states and the display will never get
turned back on.