improve drm/dri initialisation code
The main motivation here was to fix the scenario where X is started without
nouveau.ko loaded, checks if kms is enabled, is told no because nouveau.ko
isn't loaded, and then later nouveau.ko gets loaded with kms enabled by
default and starts fighting with the ddx for the display.

One cool side-effect is that we're not depending on DRIScreenInit for our
DRM fd now, which means that if DRIScreenInit fails for a non-fatal reason
(like, NVIDIA's libglx being around instead of xorg's) we can still run
with full 2D acceleration.