bios: use image from PRAMIN in preference to PROM on NV50
There's at least one known case (rh#492658) where the DCB table present in
the VBIOS image from PROM is not suitable for use.  It contained all 16
entries filled, each entry valid in itself, but contradicting other entries.

The VBIOS image in PRAMIN however, still has all 16 entries filled, but the
first few entries now match what is present on the hardware, and the rest
are set as type 0xf, which we ignore.